refactor(common): de-dup MockAsyncStreamingReadRpc
